requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
replicate==1.0.7
python-dotenv==1.0.0
boto3>=1.34.0
python-multipart>=0.0.6
aiofiles>=23.0.0
httpx>=0.25.0
PyJWT>=2.8.0
